requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
Faker==24.11.0
mysqlclient==2.2.4
pillow==10.3.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
